Rename catalog-component-instance.ts to catalog_component_instance.ts
