+++ renderers/angular/src/v0_9/catalog/types.ts
@@ -16,7 +16,7 @@
 
 import {Type} from '@angular/core';
 import {Catalog, ComponentApi} from '@a2ui/web_core/v0_9';
-import {CatalogComponentInstance} from '../core/catalog-component-instance';
+import {CatalogComponentInstance} from '../core/catalog_component_instance';
 
 /**
  * Temporary type used during basic catalog schema alignment to bypass strict type checking.
