+++ renderers/angular/src/v0_9/core/catalog_component_instance.ts
